Adding roles to users
I'm triying to add roles to users on my library app, in this commit I
add a table named roles, and add my views to function with an guest user
role, author role and editorial role. The next step that I'm going to do
is add a middleware to provide the privileges to each user role.